api/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
Pillow
drf_spectacular
channels
psycopg2-binary==2.8.5